Create LychrelSearch.java
Lychrel Search: 
Input is a series of lines terminated by end-of-file. First line is the base. The 2nd value is the number in base 10.
Program takes the number in its base form. ex.) binary, hexadecimal, from base 2...up to 256. 
Then the program performs a sum of the value in its base form and its reverse value.
The program will continue summing the value and its reverse until the resulting number is a palindrome, aka, not a lychrel number. 
The program returns the number of iterations it took to reach the palindrome value and the length of the palindrome, otherwise it returns that the number took greater than 500 iterations, a candidate Lychrel number.